WebApı established with IoC container first operation.

+++ DevCampProject/Entities/Concrete/Product.cs
@@ -11,7 +11,7 @@ namespace Entities.Concrete
         public int CategoryID { get; set; }
         public string ProductName { get; set; }
         public decimal UnitPrice { get; set; }
-        public int UnitsInStock { get; set; }
-        public string QuantityPerUnit { get; set; }
+        public short UnitsInStock { get; set; }
+        public string QuantityPerUnit{ get; set; }
     }
 }
